added test for country service

+++ src/main/scala/models/Country.scala
@@ -1,3 +1,3 @@
 package models
 
-case class Country(code: String, name: String, pop: Int, gnp: Option[Double])
+case class Country(code: String, name: String, pop: Int, gnp: Option[BigDecimal])
